\name{isa.about}

\alias{isa.about}

\title{
The library ispaadis: summary information
}

\description{
A collection of functions for analysing InterSpread Plus and AADIS simulation outputs.
}

\usage{
isa.about()
}

\details{
The most recent version of the ispaadis package can be obtained from: \url{https://fvas.unimelb.edu.au/research/groups/veterinary-epidemiology-melbourne}.
}

\author{
Mark Stevenson (University of Melbourne, Australia), Simon Firestone (University of Melbourne, Australia), Clare Death (Department of Agriculture and Water Resources, Australia), Richard Bradhurst (University of Melbourne, Australia), Robert Sanson (AsureQuality, New Zealand), Graeme Garner (Department of Agriculture and Water Resources, Australia) and Matthew Edwards (EpiVets New Zealand).
}
